# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: D&D Adventure Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /game/action:
    post:
      operationId: performAction
      tags: [game]
      summary: Perform a player action (streams DM response)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActionRequest"
      responses:
        "200":
          description: SSE stream of DM narrative
          content:
            text/event-stream: {}
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not an active player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ActionRequest:
      type: object
      properties:
        username:
          type: string
        action:
          type: string
      required:
        - username
        - action
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...